chore(tsconfig): add paths settings and tsconfig-paths dependency
Adds "baseUrl" and "paths" settings to tsconfig.json to make it easier to import modules with
absolute paths. Additionally, it adds the "tsconfig-paths" dependency to ensure that paths are
resolved correctly during project execution. Updates the "dev" script in package.json to use the
tsconfig-paths registry when starting the development server.